Error messaging, ProgressBars and No-Manifest pipelines
Implement error handling for PathComponents which collects thrown errors, provides visual feedback, and allows users to see details on where failures occur with more detail.
Implement a ProgressBar display for when a Pipeline is executed to better communicate ThunderKit activity during pipeline execution.

Update Pipelines to be allowed to execute without a manifest if there are no jobs which require or are directly using Manifests.
This will not prevent the execution of PipelineJobs using a PathReference which requires a Manifest, however the improved error handling allows users to catch this more easily.