Markdown Renderer improvements
Made improvemetns to the layouting of Markdown.
Spacing between elements was improved by normalizing how spacing is applied.
In order to support this change, LinkInlineRenderer no longer produces a wrapper element and instead directly creates its label element and applies the event handlers to the label instead.

MarkdownElement now applies the Markdown styles to itself to ensure that the markdown styles are always applied to MarkdownElements, preventing a repeated mistake in deploying MarkdownElement into a new window.
Cleaned up other locations that were applying the Markdown style as it is no longer necessary.