Make additional improvements to debugging features
The Log entry Context has been split out from the Log window to provide a more manageable and flexible interface.
Made several improvements to exception flow and logging to improve clarity of events in a pipeline run.
Stop using 2 fields for Date/Time and instead use a single DateTime field in the Log view window.
Setup the Log window to display in a compact format rather than a cozy format as the information density doesn't warrant a cozy format.