Fix Memory Leak and Text Loss
UIElementRendere was creating excessive references to MarkdownElements in its operation, this was a memory leak and additionally lead to a processing speed issue.

Fixes an issue in UIElementRenderer that would cause some text to be lost under some circumstances.  This change removes the Regex parser for finding white space and replaces it with a moving search for the space character exclusively to split literal strings.

Remove Unsupported TaskListRenderer from default renderers