Add JsonFrontMatter extension to Markdown engine
Adding a JsonFrontMatter extension to the Markdown Engine in a static form.  Will consider configurable design in the future for advanced scenarios.
The JsonFrontMatter in its current form will allow the configuration of a MarkdownElement to some extent from the FrontMatter of a Markdown document.
This allows the author to enable a header on the document with custom classes applied to its elements.
The header is a container, which contains an icon and a label.
A USS style sheet can be added to the MarkdownElement from the FrontMatter
Classes can be added to the header container, Icon, or label from the FrontMatter as well.