improve compatibility and design of enhanced import
Select extension for native assemblies based upon editor OS environment

Utilize internal Combine to further ensure standard operation inside Unity

Move Assembly blacklist generation into method an revert to building the list off of in memory assemblies which continues to appear to be the most accurate source.

If the game assembly package is already in the project, add all assemblies in that package to a whitelist.  this ensures that game assemblies will always be reimported.