add GuidMode for UTF8 in addition to default
GuidMode.Original uses the original algorithm that ThunderKit used to produce GUIDs for assemblies, this means it calls Encoding.Default which will not always produce the same encoding on different systems, limiting the cross-system compatibility it intends to create.
GuidMode.Stabilized is added and is the default setting, this mode uses Encoding.UTF8 and otherwise the same algorithm as GuidMode.Original, this should frequently produce the same GUIDs as GuidMode.Original, but is provided separate to allow for projects that do break to continue working using the old method.
GuidMode.AssetRipperCompatibility produces GUIDs using the same algorithm as AssetRipper, which is mostly the same as ThunderKit but produces GUIDs that conform to spec.