Fully implement Import Extensions
Provide Import extensions for 4 aspects of the game import process.
BlacklistProcessor and WhitelistProcessor allow the improt process to influence what assemblies get imported during the import process
AssemblyProcessor receives an assembly path for each assembly to be imported and can return a different path in order to change what is imported.  This allows for processing the assembly with publicizing, stripping and other types of patching processes during the import of game assemblies.
ConfigureAction is a custom action that occurs after all standard import steps are completed, but before the Package manifest is generated and project refreshed.