Improve ThunderstorePackageInstallater importers
ThunderstorePackageInstallers suffered from requiring a concrete ThunderstoreSource asset created in the project leading to issues with inconsistent behavior.
This update fixes that by migrating the system to use a transient ThunderstoreSource, and having implementors of ThunderstorePackageInstaller provide the URL to the store rather than the path to the asset.
This also improves how the ThunderKit Extension store is added to the project, avoiding the creation of duplicate assets.