Fix issue with Addressable loading
This solution isn't ideal.  It causes the domain reload to be exceptionally long.
A better solution would be to determine why addressables tries to reload asset bundles and prevent it or remedy the condition that leads to it.
In the mean time this is far more stable than not having it.