Update FixMisreferencedAssets.cs
Asset Reference Migration did not seem to work when an asset could not load as an object. This seemed to cause problems when attempting to get the asset path of an asset that hasn't been able to load.
Switched the system to use asset GUID's instead of Objects for getting asset paths.